Adding PUT endpoint to allow updating of prices via the REST api.

+++ code/src/main/java/com/myretail/api/model/ProductInfo.java
@@ -5,43 +5,43 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ProductInfo {
 
-    private Long id;
-    private String name;
-    private Price current_price;
+  private int id;
+  private String name;
+  private Price current_price;
 
-    public ProductInfo() {
-    }
+  public ProductInfo() {
+  }
 
-    public Long getId() {
-        return id;
-    }
+  public int getId() {
+    return id;
+  }
 
-    public void setId(Long id) {
-        this.id = id;
-    }
+  public void setId(int id) {
+    this.id = id;
+  }
 
-    public String getName() {
-        return name;
-    }
+  public String getName() {
+    return name;
+  }
 
-    public void setName(String name) {
-        this.name = name;
-    }
+  public void setName(String name) {
+    this.name = name;
+  }
 
-    public Price getCurrent_price() {
-        return current_price;
-    }
+  public Price getCurrent_price() {
+    return current_price;
+  }
 
-    public void setCurrent_price(Price current_price) {
-        this.current_price = current_price;
-    }
+  public void setCurrent_price(Price current_price) {
+    this.current_price = current_price;
+  }
 
-    @Override
-    public String toString() {
-        return "ProductInfo{" +
-                "id=" + id +
-                ", name=" + name +
-                ", current_price=" + current_price +
-                '}';
-    }
+  @Override
+  public String toString() {
+    return "ProductInfo{" +
+              "id=" + id +
+              ", name=" + name +
+              ", current_price=" + current_price +
+              "}";
+  }
 }
